requirements.txt
Django==1.5.4
lxml==3.2.3
httplib2==0.8
restclient==0.11.0
feedparser==4.1
Markdown==2.3.1
pyenchant==1.5.1
simplejson==3.4.0
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.1
statsd==2.0.2
six==1.3.0
python-dateutil==2.1
mimeparse==0.1.3
django-tastypie==0.9.15
unittest2==0.5.1
nose==1.3.0
coverage==3.5.3
BeautifulSoup==3.2.1
cssselect==0.8
fuzzywuzzy==0.1
sure==1.1.7
ipython==0.13.1
ipdb==0.7
lettuce==0.2.14
selenium==2.29.0
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
pep8==1.4
pyflakes==0.5.0
flake8==1.7.0
Pillow==1.7.8
factory_boy==2.1.1

django-smartif==0.1
python-ldap==2.4.13
djangowind==0.11.0
sorl==3.1
tagging==0.3-pre
template_utils==0.4p2
typogrify==1.0
django-staticmedia==0.2
django-indexer==0.2.1
django-templatetag-sugar==0.1
django-paging==0.2.2
django-munin==0.1.1
django-annoying==0.7.6
django-treebeard==2.0b2
South==0.8.1
django-statsd-mozilla==0.3.8.7
django-nose==1.2
django-bootstrap-form==2.0.3
django-markwhat==1.0
django-pagetree==0.14.2
django-pageblocks==0.5.11
django-quizblock==0.6.14
django-pagetree_export==0.3
django-careermapblock==0.2.16
django-registration==1.0
django-responseblock==0.1.0
django-jenkins==0.14.0
django-stagingcontext==0.1.0
django-smoketest==0.1.3
django-waffle==0.9.0 
django-sentry==1.5.4
raven==4.0.3
django-appconf==0.4.1
django_compressor==1.3
